The spaghetti code itself
This similarity of the game was created by me several years ago. At that time, I studied the basics of programming. Criticism and other is not relevant. Feel free to take the idea if interesting, copy and distribute. I added it for history, because I lost a lot of sources.